Updates web classes to accept new files and cellular_compartment
The web sub classes need to accept a cellular compartment, so that they
can change the results based on the chosen cellular compartment. They
also need to be able to accept provided true positive and false positive
files, if the cellular compartment chosen is not one that is pre-cached
from uniprot.