Fixes handling of '#N/A' in web user input.
The webuserinputreader and the cliuserinputreader were using the
pd.DataFrame and pd.read_table methods of pandas to load in the mass
spec data provided by the user. The pd.read_table method converts the
'#N/A' string into an np.nan during the loading, but the pd.DataFrame
method does not. This meant that uploading the same file to both the cli
and the website would result in the correct output on the cli and a
failure on website. This fix converts webuserinputreader to use the
pd.read_table method so that both the cli and the web interface work
with the uploaded file unchanged.