Adds an 'Entry' header to all upload tp/fp files
The uploaded true positive and false positive files need to have a
header with the text 'Entry' in order for the current code to work. The
code to add a header was already in the command line version of the
processor. This update adds it the web version so that they function
the same.